Created separate js file for translations
